+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 
 ### Fixed
 
-Writes debug messages as verbose messages.
+* Writes debug messages as verbose messages.
+* `Invoke-KernelGetNumaHighestNodeNumber` always returns `1`.
 
 ## 2.0.0