+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 
 # PureInvoke PowerShell Module Changelog
 
+## 2.0.1
+
+### Fixed
+
+* Writes debug messages as verbose messages.
+* `Invoke-KernelGetNumaHighestNodeNumber` always returns `1`.
+
 ## 2.0.0
 
 ### Upgrade Instructions